% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPropensity.R
\name{ExpectedLabelPairs}
\alias{ExpectedLabelPairs}
\title{ExpectedLabelPairs}
\usage{
ExpectedLabelPairs(object, slot, label.list = NULL, reduction = "pca")
}
\arguments{
\item{object}{The input Seurat object}

\item{slot}{The slot in object@meta.data containing the labels assigned to the cells}

\item{label.list}{external list of labels to consider (default: all labels in the metadata slot)}
}
\value{
The expected matrix
}
\description{
Compute the matrix containing the expected count of paired labels, given the number of label occurrences and the number of the neighbours
}
