% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPropensity.R
\name{PrepareSeuratObject}
\alias{PrepareSeuratObject}
\title{PrepareSeuratObject}
\usage{
PrepareSeuratObject(
  object,
  cells = NULL,
  reduction = "",
  dims = 10,
  k = 30,
  ...
)
}
\arguments{
\item{object}{The input Seurat object}

\item{cells}{The cell ID to subset the object on (default: use all cells)}

\item{reduction}{The pre-computed dimensional reduction to use (default: run PCA)}

\item{dims}{The number of top dimensions to use (default: 10)}

\item{k}{The number of neighbours in the k-nn graph (if not already computed)}

\item{selection.method}{The method to use for computing HVGs (default: use the already computed hvgs)}
}
\value{
The object with updated hvg, PC, and k-nn
}
\description{
Compute highly variable genes, PCA and k-nn graph
}
