% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPropensity.R
\name{Propensity}
\alias{Propensity}
\title{Propensity}
\usage{
Propensity(object, slot, label.list = NULL, reduction = "pca")
}
\arguments{
\item{object}{The input Seurat object}

\item{slot}{The slot in object@meta.data containing the labels assigned to the cells}

\item{label.list}{external list of labels to consider (default: all labels in the metadata slot)}

\item{reduction}{dimensional reduction (default: "pca")}
}
\value{
The propensity matrix
}
\description{
Compute the pair propensity matrix
}
