% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPropensity.R
\name{Symmetrise}
\alias{Symmetrise}
\title{Symmetrise}
\usage{
Symmetrise(C)
}
\arguments{
\item{C}{The propensity matrix}
}
\value{
The simmetric propensity matrix
}
\description{
Compute the symmetric version of the pair propensity matrix (L2 norm)
}
