% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPropensity.R
\name{WriteMatrix}
\alias{WriteMatrix}
\title{WriteMatrix}
\usage{
WriteMatrix(C, filename)
}
\arguments{
\item{C}{The propensity matrix}

\item{filename}{tsv filename where to print the matrix}
}
\value{
The simmetric propensity matrix
}
\description{
Write output propensity matrices (before and after L2 normalisation
}
